# api/openapi-v1.yaml
openapi: 3.0.0
info:
  title: Briany BPM API
  version: 2.0.0
  description: |
    REST API for a Flowable-based BPM platform.

    The target contract after the REST redesign. The spec deliberately separates four
    audiences:
    - **public** - business and workflow operations for client applications;
    - **admin** - operator and observability endpoints;
    - **internal** - dev/debug tooling, not for external clients;
    - **testing** - e2e and destructive test utilities.

    All list endpoints return paginated results in a unified `PagedResponse` shape
    (`page`, `pageSize`, `totalElements`, `totalPages`, `hasNext`, `hasPrevious`).

    Authentication is HTTP Basic (Flowable IDM) or Bearer JWT (JWKS) - see the
    Security section.

  contact:
    email: ivan@moarse.ru
    name: Ivan Donchenko
  license:
    name: ''
servers:
  - url: 'http://10.6.18.196:8095'
    description: Developer Stand
security:
  - BasicAuth: []
  - BearerAuth: []
...
paths:
  /api/v1/applications:
    get:
      tags:
        - Application
      x-audience: public
      x-spring-paginated: true
      x-controller-replaces:
        - work/AppDefinitionsController
      operationId: listApplications
      summary: List applications (latest versions)
      description: |
        Returns a paginated list of workflow applications.
        Only the latest version of each application key is returned.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Page of applications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      tags:
        - Application
      x-audience: public
      x-controller-replaces:
        - deployment/DeploymentController
      operationId: deployApplication
      summary: Deploys an application from a ZIP archive
      description: |
        Uploads a ZIP archive containing BPMN, DMN, bform, and other resources.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: ZIP archive containing bpmn/dmn/app/bform files.
      responses:
        '201':
          description: Application deployed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/applications/versions:
    get:
      tags:
        - Application
      x-audience: public
      x-spring-paginated: true
      x-controller-replaces:
        - work/AppDefinitionsController
      operationId: listApplicationsVersions
      summary: List applications (all versions)
      description: |
        Returns a paginated list of workflow applications across all deployed versions.
        Each version is returned as a separate entry.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Page of applications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  '/api/v1/applications/{key}':
    get:
      tags:
        - Application
      x-audience: public
      x-controller-replaces:
        - work/AppDefinitionsController
      operationId: getApplication
      summary: Get application by key (latest version)
      description: |
        Returns a workflow application by its id or key.
        If multiple versions are deployed, only the latest is returned.
        Use `getApplicationVersion` for a specific version.
      responses:
        '200':
          description: Application data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    parameters:
      - $ref: '#/components/parameters/Key'
  '/api/v1/applications/{key}/versions':
    get:
      tags:
        - Application
      x-audience: public
      x-spring-paginated: true
      x-controller-replaces:
        - work/AppDefinitionsController
      operationId: listApplicationVersions
      summary: Get application by key (all versions)
      description: |
        Returns all versions of an application by key or id.
        Unlike `getApplication`, not limited to the latest version.
      responses:
        '200':
          description: Page of application versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    parameters:
      - $ref: '#/components/parameters/Key'
  '/api/v1/applications/{key}/versions/{version}':
    get:
      tags:
        - Application
      x-audience: public
      x-controller-replaces:
        - work/AppDefinitionsController
      operationId: getApplicationVersion
      summary: Get application by id or key (specific version)
      description: |
        Returns a specific application version by key or id and version number.
      responses:
        '200':
          description: Application data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    parameters:
      - $ref: '#/components/parameters/Key'
      - $ref: '#/components/parameters/Version'
  /api/v1/processes:
    get:
      tags:
        - ProcessDefinition
      x-audience: public
      x-spring-paginated: true
      x-controller-replaces:
        - work/ProcessDefinitionsController
      x-with-principal: true
      operationId: listProcesses
      summary: List process definitions (latest versions)
      description: |
        Returns a paginated list of all process definitions.
        By default only the latest version of each definition is returned.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
      responses:
        '200':
          description: Page of process definitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessDefinitionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  '/api/v1/processes/{key}':
    get:
      tags:
        - ProcessDefinition
      x-audience: public
      x-controller-replaces:
        - work/ProcessDefinitionsController
      x-with-principal: true
      operationId: getProcess
      summary: Get process definition by key (latest version)
      description: |
        Returns the latest process definition for a key.
      responses:
        '200':
          description: Process definition data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessDefinition'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    parameters:
      - $ref: '#/components/parameters/Key'
  '/api/v1/processes/{key}/versions':
    get:
      tags:
        - ProcessDefinition
      x-audience: public
      x-spring-paginated: true
      x-controller-replaces:
        - work/ProcessDefinitionsController
      x-with-principal: true
      operationId: listProcessDefinitionVersions
      summary: Get all versions of a process definition
      description: |
        Returns all versions of a process definition by key or id.
        Unlike `getProcessDefinition`, not limited to the latest version.
      responses:
        '200':
          description: Page of process definitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessDefinitionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    parameters:
      - $ref: '#/components/parameters/Key'
  '/api/v1/processes/{key}/versions/{version}':
    get:
      tags:
        - ProcessDefinition
      x-audience: public
      x-controller-replaces:
        - work/ProcessDefinitionsController
      x-with-principal: true
      operationId: getProcessDefinitionVersion
      summary: Get process definition by version
      description: |
        Returns a specific process definition version by key or id and version number.
      responses:
        '200':
          description: Process definition data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessDefinition'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    parameters:
      - $ref: '#/components/parameters/Key'
      - $ref: '#/components/parameters/Version'
  '/api/v1/processes/{key}/process-instances':
    get:
      tags:
        - ProcessInstance
      x-spring-paginated: true
      x-audience: public
      x-controller-replaces:
        - work/ProcessInstanceQueryController
      x-with-principal: true
      operationId: listProcessProcessInstances
      summary: List instances of a process definition
      description: |
        Returns a paginated list of process instances started from the given process
        definition (across all its versions).
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/ProcessInstanceState'
      responses:
        '200':
          description: Page of process instances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessInstancePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    parameters:
      - $ref: '#/components/parameters/Key'
  '/api/v1/processes/{key}/versions/{version}/process-instances':
    get:
      tags:
        - ProcessInstance
      x-spring-paginated: true
      x-audience: public
      x-controller-replaces:
        - work/ProcessInstanceQueryController
      x-with-principal: true
      operationId: listProcessDefinitionVersionInstances
      summary: List instances of a single process definition version
      description: |
        Returns process instances for one specific definition version.
        Unlike `listProcessDefinitionInstances`, filters only by the given version.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/ProcessInstanceState'
      responses:
        '200':
          description: Page of process instances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessInstancePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    parameters:
      - $ref: '#/components/parameters/Key'
      - $ref: '#/components/parameters/Version'
  /api/v1/process-instances:
    get:
      tags:
        - ProcessInstance
      x-audience: public
      x-spring-paginated: true
      x-controller-replaces:
        - work/ProcessInstanceQueryController
      x-with-principal: true
      operationId: listProcessInstances
      summary: List process instances
      description: |
        Returns a paginated list of all process instances.
        By default both active and completed instances are returned.
      parameters:
        - $ref: '#/components/parameters/SortParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/ProcessInstanceState'
      responses:
        '200':
          description: Page of process instances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessInstancePage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: Checked against Flowable IDM (see FlowableAuthStrategy).
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer JWT validated against a JWKS endpoint (see JwksStrategy).
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 25
    SortParam:
      name: sort
      in: query
      description: 'Sort criterion: field[,asc|desc]. Example: name,asc'
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
    Key:
      name: key
      in: path
      required: true
      schema:
        type: string
      description: definition key
    ProcessInstanceState:
      name: state
      in: query
      schema:
        enum:
          - running
          - completed
          - suspended
      description: |
        Filters by process instance state. If omitted, all processes are returned.
        - `running` - not yet finished (active and suspended)
        - `completed` - finished
    Version:
      name: version
      in: path
      required: true
      schema:
        type: integer
      description: Version identifier
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentication is required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated principal is not allowed to access the resource
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: State conflict or business rule violation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailable:
      description: Upstream dependency is unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required:
        - detail
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        instance:
          type: string
          format: uri-reference
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
    AppDefinition:
      type: object
      example:
        id: 'app:12'
        key: loans
        version: 1
        name: Loans Workspace
        description: Loan workflow cockpit
        icon: bank
        theme: default
      required:
        - id
        - key
        - name
        - version
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        version:
          type: integer
        resourceName:
          type: string
        deploymentId:
          type: string
    Application:
      example:
        id: 'app:12'
        key: loans
        version: 1
        name: Loans Workspace
        description: Loan workflow cockpit
        icon: bank
        theme: default
      allOf:
        - type: object
          properties:
            deployedResources:
              type: object
              properties:
                processDefinitions:
                  type: array
                  items:
                    $ref: '#/components/schemas/ResourceRef'
                bforms:
                  type: array
                  items:
                    $ref: '#/components/schemas/ResourceRef'
        - $ref: '#/components/schemas/ApplicationRef'
    ApplicationRef:
      allOf:
        - type: object
          properties:
            icon:
              type: string
            theme:
              type: string
            deploymentTime:
              type: string
              format: date-time
        - $ref: '#/components/schemas/AppDefinition'
    ApplicationPage:
      x-kotlin-implements:
        - ru.briany.common.api.dtos.PagedList<ApplicationRef>
      x-kotlin-implements-fields:
        - data
        - page
        - pageSize
        - totalElements
        - totalPages
        - hasNext
        - hasPrevious
      allOf:
        - type: object
          required:
            - data
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ApplicationRef'
        - $ref: '#/components/schemas/PagedResponse'
    ProcessDefinition:
      type: object
      example:
        id: 'orderProcess:7:4a0716d3'
        key: orderProcess
        name: Order Process
        description: Order approval and fulfillment process
        version: 7
        category: workflow
        deploymentId: dep-7741
        hasStartForm: false
      required:
        - id
        - key
        - version
        - deploymentId
        - hasStartForm
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: integer
        category:
          type: string
        deploymentId:
          type: string
        hasStartForm:
          type: boolean
          default: false
    ProcessDefinitionPage:
      x-kotlin-implements:
        - ru.briany.common.api.dtos.PagedList<ProcessDefinition>
      x-kotlin-implements-fields:
        - data
        - page
        - pageSize
        - totalElements
        - totalPages
        - hasNext
        - hasPrevious
      allOf:
        - type: object
          required:
            - data
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ProcessDefinition'
        - $ref: '#/components/schemas/PagedResponse'
    ResourceRef:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        key:
          type: string
        name:
          type: string
        version:
          type: integer
      title: Short resource reference
      description: 'Used in compact collection listings (bpmn, cmmn, dmn, bform)'
    ProcessInstance:
      type: object
      example:
        id: '2501'
        name: Order
        businessKey: ORD-2026-0001
        state: running
        startedAt: '2026-04-07T09:30:00Z'
        startedBy: '1001'
        processDefinition:
          id: 'orderProcess:7:4a0716d3'
          key: orderProcess
          name: Order Process
          version: 7
        graphicalNotationDefined: true
      required:
        - id
        - state
      properties:
        id:
          type: string
        name:
          type: string
        businessKey:
          type: string
        state:
          $ref: '#/components/schemas/ProcessInstanceState'
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        startUserId:
          type: string
        durationInMillis:
          type: integer
          format: int64
        processDefinition:
          $ref: '#/components/schemas/ResourceRef'
    ProcessInstancePage:
      x-kotlin-implements:
        - ru.briany.common.api.dtos.PagedList<ProcessInstance>
      x-kotlin-implements-fields:
        - data
        - page
        - pageSize
        - totalElements
        - totalPages
        - hasNext
        - hasPrevious
      allOf:
        - type: object
          required:
            - data
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ProcessInstance'
        - $ref: '#/components/schemas/PagedResponse'
...
    PagedResponse:
      title: PagedResponse
      type: object
      required:
        - data
        - page
        - pageSize
        - totalElements
        - totalPages
        - hasNext
        - hasPrevious
      properties:
        data:
          type: array
          items:
            type: object
        page:
          type: integer
        pageSize:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    ProcessInstanceState:
      title: ProcessInstanceState
      enum:
        - running
        - completed
        - suspended
...